+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Example debug command and attach configurations for C++, Python, and Java as well as convenient ways to automatically add them to launch.json file
 - Toggle show details button for stress tester as well as more statuses for judge
 - Icon to open the corresponding file for stress tester
+- Icon to open the interactor for judge
 
 ### Changed